attept to fix flakey startup error
