+++ lib/fmt/call.go
@@ -7,10 +7,6 @@ import (
 	"io"
 	"net/http"
 	"os"
-	"strings"
-	"time"
-
-	"github.com/hofstadter-io/hof/lib/container"
 )
 
 // this file has functions for
@@ -73,78 +69,3 @@ func (fmtr *Formatter) Call(filename string, content []byte, config any) ([]byte
 	return content, nil
 }
 
-func (fmtr *Formatter) WaitForRunning(retry int, delay time.Duration) error {
-
-	// We should probably rethink how this works, such that
-	// we minimize our exec out to docker (et al)
-	// we can exec once and then check on all formatters
-
-	// fmt.Println("wait-running.0:", fmtr.Name, fmtr.Status, fmtr.Running, fmtr.Ready)
-	// return if already running
-	if fmtr.Running {
-		return nil
-	}
-
-	for i := 0; i < retry; i++ {
-		containers, err := container.GetContainers(ContainerPrefix + fmtr.Name)
-		if err != nil {
-			return err
-		}
-
-		for _, container := range containers {
-			// extract name
-			name := container.Names[0]
-			name = strings.TrimPrefix(name, ContainerPrefix)
-			// fmt.Println("wait-running:", fmtr.Name, name, container.State)
-			if name == fmtr.Name {
-				c := container
-				updateFmtrContainer(fmtr, &c)
-				break
-			}
-		}
-
-		if fmtr.Running {
-			return nil
-		}
-
-		time.Sleep(delay)
-	}
-
-	return fmt.Errorf("formatter %s never started", fmtr.Name)
-}
-
-func (fmtr *Formatter) WaitForReady(retry int, delay time.Duration) error {
-	// fmt.Println("wait-ready.0:", fmtr.Name, fmtr.Status, fmtr.Running, fmtr.Ready)
-
-	// return if already ready
-	if fmtr.Ready {
-		return nil
-	}
-
-	// return error if not running
-	if !fmtr.Running {
-		return fmt.Errorf("formatter %s is not running", fmtr.Name)
-	}
-
-	// get ready check payload
-	p, ok := fmtrReady[fmtr.Name]
-	if !ok {
-		fmt.Printf("warn: formatter %s does not have a ready config\n", fmtr.Name)
-		return nil
-	}
-
-	payload := p.(map[string]any)
-
-	for i := 0; i < retry; i++ {
-		_, err := fmtr.Call("ready-check", []byte(payload["source"].(string)), payload["config"])
-		fmt.Println("wait-ready:", i, fmtr.Name, fmtr.Ready, err)
-		// if no error, then ready
-		if err == nil {
-			fmtr.Ready = true
-			return nil
-		}
-		time.Sleep(delay)
-	}
-
-	return fmt.Errorf("formatter %s is not ready", fmtr.Name)
-}
