+++ lib/fmt/status.go
@@ -3,6 +3,8 @@ package fmt
 import (
 	"fmt"
 	"strings"
+	"time"
+
 	"github.com/hofstadter-io/hof/lib/container"
 )
 
@@ -60,6 +62,26 @@ func UpdateFormatterStatus() error {
 	return nil
 }
 
+func getAndUpdateFmtrContainer(fmtr *Formatter) error {
+	containers, err := container.GetContainers(ContainerPrefix + fmtr.Name)
+	if err != nil {
+		return err
+	}
+
+	for _, c := range containers {
+		// extract name
+		name := c.Names[0]
+		name = strings.TrimPrefix(name, ContainerPrefix)
+		if name == fmtr.Name {
+			C := c
+			updateFmtrContainer(fmtr, &C)
+			break
+		}
+	}
+
+	return nil
+}
+
 func updateFmtrContainer(fmtr *Formatter, container *container.Container) {
 
 		fmtr.Status = container.State
@@ -88,3 +110,71 @@ func updateFmtrContainer(fmtr *Formatter, container *container.Container) {
 		fmtr.Container = container
 }
 
+func (fmtr *Formatter) WaitForRunning(retry int, delay time.Duration) error {
+
+	// We should probably rethink how this works, such that
+	// we minimize our exec out to docker (et al)
+	// we can exec once and then check on all formatters
+
+	// fmt.Println("wait-running.0:", fmtr.Name, fmtr.Status, fmtr.Running, fmtr.Ready)
+	// return if already running
+	if fmtr.Running {
+		return nil
+	}
+
+	for i := 0; i < retry; i++ {
+
+		err := getAndUpdateFmtrContainer(fmtr)
+		if err != nil {
+			return err
+		}
+
+		if fmtr.Running {
+			return nil
+		}
+
+		time.Sleep(delay)
+	}
+
+	return fmt.Errorf("formatter %s never started", fmtr.Name)
+}
+
+func (fmtr *Formatter) WaitForReady(retry int, delay time.Duration) error {
+	// fmt.Println("wait-ready.0:", fmtr.Name, fmtr.Status, fmtr.Running, fmtr.Ready)
+	err := getAndUpdateFmtrContainer(fmtr)
+	if err != nil {
+		return err
+	}
+
+	// return if already ready
+	if fmtr.Ready {
+		return nil
+	}
+
+	// return error if not running
+	if !fmtr.Running {
+		return fmt.Errorf("formatter %s is not running", fmtr.Name)
+	}
+
+	// get ready check payload
+	p, ok := fmtrReady[fmtr.Name]
+	if !ok {
+		fmt.Printf("warn: formatter %s does not have a ready config\n", fmtr.Name)
+		return nil
+	}
+
+	payload := p.(map[string]any)
+
+	for i := 0; i < retry; i++ {
+		_, err := fmtr.Call("ready-check", []byte(payload["source"].(string)), payload["config"])
+		// fmt.Println("wait-ready:", i, fmtr.Name, fmtr.Ready, err)
+		// if no error, then ready
+		if err == nil {
+			fmtr.Ready = true
+			return nil
+		}
+		time.Sleep(delay)
+	}
+
+	return fmt.Errorf("formatter %s is not ready", fmtr.Name)
+}
